requirements.txt
tweepy
python-dotenv
feedparser
requests
beautifulsoup4
google-generativeai
